+++ app/controllers/settings_controller.rb
@@ -51,6 +51,7 @@ class SettingsController < ApplicationController
         params.require( :setting ).
             permit( :scan_global_limit, :scan_per_user_limit, { scan_allowed_types: [] },
                     { profile_allowed_modules: [] }, { profile_allowed_plugins: [] },
-                    :scan_target_whitelist_patterns, :scan_target_blacklist_patterns )
+                    :scan_target_whitelist_patterns, :scan_target_blacklist_patterns,
+                    :timezone )
     end
 end