+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20131010013423) do
+ActiveRecord::Schema.define(version: 20131010190108) do
 
   create_table "comments", force: true do |t|
     t.integer  "user_id"
@@ -235,6 +235,7 @@ ActiveRecord::Schema.define(version: 20131010013423) do
     t.datetime "created_at"
     t.datetime "updated_at"
     t.boolean  "welcomed"
+    t.string   "timezone"
   end
 
   create_table "users", force: true do |t|