+++ CHANGELOG.md
@@ -22,6 +22,8 @@
 - Settings
     - Profiles
         - Fixed heading (_Profiles_ => _Plugins_).
+    - Added "General" tab.
+        - Added Timezone setting.
 
 ## 0.4.2.1 _(September 18, 2013)_